refactor : auth와 user 패키지를 합침
