--- com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/auth/domain/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.sparta.cupeed.user.auth.domain.repository;
-
-import java.util.Optional;
-
-import com.sparta.cupeed.user.auth.domain.model.User;
-import com.sparta.cupeed.user.auth.domain.model.UserCompany;
-
-public interface UserRepository {
-
-	Optional<User> findByUserId(String userId);
-
-	Optional<User> findBySlackId(String slackId);
-
-	void save(User user, UserCompany newUserCompany);
-}
