--- com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/auth/domain/vo/UserRoleEnum.java
@@ -1,10 +1,0 @@
-package com.sparta.cupeed.user.auth.domain.vo;
-
-public enum UserRoleEnum {
-
-	MASTER, HUB, DELIVERY, COMPANY;
-
-	public String getAuthority() {
-		return "ROLE_" + this.name();
-	}
-}
