--- com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/auth/domain/vo/UserStatusEnum.java
@@ -1,6 +1,0 @@
-package com.sparta.cupeed.user.auth.domain.vo;
-
-public enum UserStatusEnum {
-
-	PENDING, APPROVED, REJECTED
-}
