--- com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/auth/presentation/advice/AuthException.java
@@ -1,11 +1,0 @@
-package com.sparta.cupeed.user.auth.presentation.advice;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class AuthException extends RuntimeException {
-
-	private final AuthError authError;
-}
