+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/domain/vo/UserRoleEnum.java
@@ -1,0 +1,10 @@
+package com.sparta.cupeed.user.domain.vo;
+
+public enum UserRoleEnum {
+
+	MASTER, HUB, DELIVERY, COMPANY;
+
+	public String getAuthority() {
+		return "ROLE_" + this.name();
+	}
+}
