+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/domain/vo/UserStatusEnum.java
@@ -1,0 +1,6 @@
+package com.sparta.cupeed.user.domain.vo;
+
+public enum UserStatusEnum {
+
+	PENDING, APPROVED, REJECTED
+}
