+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/infrastructure/security/auth/UserDetailsImpl.java
@@ -1,0 +1,48 @@
+package com.sparta.cupeed.user.infrastructure.security.auth;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.UUID;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import com.auth0.jwt.interfaces.DecodedJWT;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserDetailsImpl implements UserDetails {
+
+	private UUID id;
+	private String userId;
+	private String role;
+
+	public static UserDetailsImpl of(DecodedJWT token) {
+		return UserDetailsImpl.builder()
+			.id(UUID.fromString(token.getClaim("id").asString()))
+			.userId(token.getSubject())
+			.role(token.getClaim("role").asString())
+			.build();
+	}
+
+	@Override
+	public Collection<? extends GrantedAuthority> getAuthorities() {
+		SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(role);
+
+		return List.of(simpleGrantedAuthority);
+	}
+
+	@Override
+	public String getPassword() {
+		return null;
+	}
+
+	@Override
+	public String getUsername() {
+		return userId;
+	}
+}
