+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/infrastructure/security/config/PasswordConfig.java
@@ -1,0 +1,15 @@
+package com.sparta.cupeed.user.infrastructure.security.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class PasswordConfig {
+
+	@Bean
+	public PasswordEncoder passwordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+}
