+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/presentation/advice/AuthException.java
@@ -1,0 +1,11 @@
+package com.sparta.cupeed.user.presentation.advice;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class AuthException extends RuntimeException {
+
+	private final AuthError authError;
+}
