+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/presentation/controller/UserControllerV1.java
@@ -1,0 +1,12 @@
+package com.sparta.cupeed.user.presentation.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequestMapping("/v1/user")
+@RequiredArgsConstructor
+public class UserControllerV1 {
+}
