--- com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/user/controller/UserController.java
@@ -1,12 +1,0 @@
-package com.sparta.cupeed.user.user.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequestMapping("/v1/user")
-@RequiredArgsConstructor
-public class UserController {
-}
